requirements.txt
langchain
openai==0.28
huggingface_hub
python-dotenv
streamlit